[ALPS04791443] v4l2: Fixed input driven ts
need codec driver update ts from bs to fm
& kernel get fm ts from vsi

MTK-Commit-Id: ad0049e1de6eba8c92e33687c7c62e73f0121224

CR-Id: ALPS04791443
Feature: H.264 Decoder

+++ drivers/media/platform/mtk-vcodec/mtk_vcodec_drv.h
@@ -165,7 +165,7 @@ struct mtk_dec_params {
 	unsigned int	nal_size_length;
 	unsigned int	svp_mode;
 	unsigned int	operating_rate;
-	unsigned int	timestamp;
+	u64	timestamp;
 	unsigned int	total_frame_bufq_count;
 	unsigned int	queued_frame_buf_count;
 };
